Update tests and test results to reflect last changes.

+++ plearn/python/test/InterfunctionXchgTest.cc
@@ -198,7 +198,7 @@ void InterfunctionXchgTest::perform()
 
 
     try {
-        map<string,int64_t> mapsd;
+        map<string,int> mapsd;
         string str_mapsd = "{ Oui:16 il:32 est:64 juste:128 et:256 bon:512 }";
         PStream is_mapsd = openString(str_mapsd, PStream::plearn_ascii);
         is_mapsd >> mapsd;
@@ -206,7 +206,7 @@ void InterfunctionXchgTest::perform()
         python_other->setGlobalObject("some_global_map", PythonObjectWrapper(mapsd));
         cout << "Associated 'some_global_map' with: " << tostring(mapsd) << endl;
         cout << "Read back from Python environment: "
-             << tostring(python_other->getGlobalObject("some_global_map").as< map<string,int64_t> >())
+             << tostring(python_other->getGlobalObject("some_global_map").as< map<string,int> >())
              << endl;
         python_other->invoke("print_global_map");
